+++ src/WebHook/Notification/MessageNotificationFactory.php
@@ -40,6 +40,8 @@ class MessageNotificationFactory
                     new Support\Business($metadata['phone_number_id'], $metadata['display_phone_number']),
                     $message[$message['type']]['id'],
                     $message[$message['type']]['mime_type'],
+                    $message[$message['type']]['sha256'],
+                    $message[$message['type']]['filename'] ?? '',
                     $message[$message['type']]['caption'] ?? '',
                     $message['timestamp']
                 );
@@ -129,7 +131,7 @@ class MessageNotificationFactory
             }
 
             $notification->withContext(new Support\Context(
-                $message['context']['id'],
+                $message['context']['id'] ?? null,
                 $message['context']['forwarded'] ?? false,
                 $referred_product ?? null
             ));